espressif32 6.5.0, AsyncTCP-esphome@^2.0.1, Dimmer float

+++ src/config/enums.h
@@ -40,7 +40,7 @@ struct DisplayValues {
   String IP;
   String time;
   bool injection; 
-  int dimmer;
+  float dimmer;
   int security; 
   int change;
   bool task;
